refactor: extract composer dispatch from AgentWorkspace
